+++ internal/app/service/service.go
@@ -4,9 +4,9 @@ import (
 	"database/sql"
 	"fmt"
 	"github.com/friendsofgo/errors"
-	"github.com/kevinmichaelchen/api-dispatch/internal/distance"
 	"github.com/kevinmichaelchen/api-dispatch/internal/service"
 	"github.com/kevinmichaelchen/api-dispatch/internal/service/db"
+	"github.com/kevinmichaelchen/api-dispatch/internal/service/distance"
 	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
 	"go.uber.org/fx"
 	"go.uber.org/zap"
